finish

+++ day01_ex00/src/User.java
@@ -20,7 +20,7 @@ public class User {
         this.name = name;
     }
 
-    public float getBalance() {
+    public int getBalance() {
         return balance;
     }
